Added methods folder, and structured archtecture
