+++ main.py
@@ -3,52 +3,15 @@ BrainFlow AI - Open Source Project
 Purpose: Use AI to process brain signals (EEG, EMG, etc.) for neuro-assistive applications.
 
 """
-
-import numpy as np
-import pandas as pd
-
-Placeholder for signal loading
-def load_brain_signal(file_path):
-    """
-    Load brain signal data from a CSV or similar format.
-    """
-    try:
-        data = pd.read_csv(file_path)
-        print(f"Loaded data shape: {data.shape}")
-        return data
-    except Exception as e:
-        print(f"Error loading data: {e}")
-        return None
-
-Placeholder for signal preprocessing
-def preprocess_signal(data):
-    """
-    Clean and preprocess the raw signal data (e.g., normalize, filter).
-    """
-    if data is None:
-        return None
-    # Dummy example: normalize the signal
-    return (data - data.mean()) / data.std()
-
-Placeholder for model training
-def train_model(processed_data):
-    """
-    Train a model on the processed brain signal data.
-    """
-    print("Training model... (placeholder)")
-    # TODO: Add model training code here (e.g., sklearn, PyTorch)
-
-Placeholder for prediction
-def predict_signal(model, new_data):
-    """
-    Use the trained model to predict based on new brain signal input.
-    """
-    print("Predicting... (placeholder)")
-    # TODO: Add prediction logic here
+from methods.process_data import load_brain_signal
+from methods.process_data import preprocess_signal
+from methods.train import train_model
+from methods.train import predict_signal
 
 Main runner
 if __name__ == "__main__":
-    file_path = "sample_data.csv"  # Replace with your dataset
+    file_path = "./data/raw/brain_signals.csv"  # Can be easily replaced with your dataset
     raw_data = load_brain_signal(file_path)
     processed = preprocess_signal(raw_data)
     train_model(processed)
+
