+++ methods/process_data.py
@@ -1,0 +1,25 @@
+import numpy as np
+import pandas as pd
+
+Placeholder for signal loading
+def load_brain_signal(file_path):
+    """
+    Load brain signal data from a CSV or similar format.
+    """
+    try:
+        data = pd.read_csv(file_path)
+        print(f"Loaded data shape: {data.shape}")
+        return data
+    except Exception as e:
+        print(f"Error loading data: {e}")
+        return None
+
+Placeholder for signal preprocessing
+def preprocess_signal(data):
+    """
+    Clean and preprocess the raw signal data (e.g., normalize, filter).
+    """
+    if data is None:
+        return None
+    # Dummy example: normalize the signal
+    return (data - data.mean()) / data.std()
