+++ methods/train.py
@@ -1,0 +1,15 @@
+Placeholder for model training
+def train_model(processed_data):
+    """
+    Train a model on the processed brain signal data.
+    """
+    print("Training model... (placeholder)")
+    # TODO: Add model training code here (e.g., sklearn, PyTorch)
+
+Placeholder for prediction
+def predict_signal(model, new_data):
+    """
+    Use the trained model to predict based on new brain signal input.
+    """
+    print("Predicting... (placeholder)")
+    # TODO: Add prediction logic here
